third_party/WebKit/LayoutTests/media/track/regions-webvtt/vtt-region-constructor.html

<!DOCTYPE html>
<title>Tests the constructor and mutation of VTTRegion.</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
test(function() {
    var region = new VTTRegion();
    assert_true(region instanceof VTTRegion, "instanceof");

    assert_equals(region.scroll, "");
    assert_equals(region.viewportAnchorX, 0);
    assert_equals(region.viewportAnchorY, 100);
    assert_equals(region.regionAnchorX, 0);
    assert_equals(region.regionAnchorY, 100);
    assert_equals(region.lines, 3);
    assert_equals(region.width, 100);

    region.scroll = "invalid-scroll-value";
    assert_equals(region.scroll, "");

    var invalidPercentageValues = [-1, 101];
    for (var value of invalidPercentageValues) {
      assert_throws("IndexSizeError", function() { region.viewportAnchorX = value; });
      assert_equals(region.viewportAnchorX, 0);
      assert_throws("IndexSizeError", function() { region.viewportAnchorY = value; });
      assert_equals(region.viewportAnchorY, 100);
      assert_throws("IndexSizeError", function() { region.regionAnchorX = value; });
      assert_equals(region.regionAnchorX, 0);
      assert_throws("IndexSizeError", function() { region.regionAnchorY = value; });
      assert_equals(region.regionAnchorY, 100);
      assert_throws("IndexSizeError", function() { region.width = value; });
      assert_equals(region.width, 100);
    }

    invalidPercentageValues = [-Infinity, Infinity, NaN];
    for (var value of invalidPercentageValues) {
      assert_throws(new TypeError, function() { region.viewportAnchorX = value; });
      assert_equals(region.viewportAnchorX, 0);
      assert_throws(new TypeError, function() { region.viewportAnchorY = value; });
      assert_equals(region.viewportAnchorY, 100);
      assert_throws(new TypeError, function() { region.regionAnchorX = value; });
      assert_equals(region.regionAnchorX, 0);
      assert_throws(new TypeError, function() { region.regionAnchorY = value; });
      assert_equals(region.regionAnchorY, 100);
      assert_throws(new TypeError, function() { region.width = value; });
      assert_equals(region.width, 100);
    }

    assert_throws("IndexSizeError", function() { region.lines = -1; });
    assert_equals(region.lines, 3);

    region.lines = 130;
    assert_equals(region.lines, 130);
    region.viewportAnchorX = 64;
    assert_equals(region.viewportAnchorX, 64);
    region.viewportAnchorY = 32;
    assert_equals(region.viewportAnchorY, 32);
    region.regionAnchorX = 16;
    assert_equals(region.regionAnchorX, 16);
    region.regionAnchorY = 8;
    assert_equals(region.regionAnchorY, 8);
    region.width = 42;
    assert_equals(region.width, 42);
});
</script>
